Syntax: Monospace Error on Empty Line
Allow monospace to carry on multiple lines, and even include comment
lines (ignored in output, and parsed correclty in syntax), but if an
empty line is encountered before closing delimiter mark the empty line
as error.

This greatly improve monospace parsing, and quotes in general.

Update monospace test accordingly, and polish its text.