Fix Broken Example Blocks
This commit fixes a bug in example blocks where curly quotes would
result in prematurely popping-out of the example context, causing the
closing block delimiter to be treated as a new opening example block,
jeopardizing syntax highlighting of the rest of the document.

This was the most annoying bug in the package, leading to documents
highlighting and scoping frequently breaking up.