Introduce Syntax Branching
In order to improve handling of common occurrences of literal `#` and
`*` characters within the text, this commit replaces the previous
fallback on empty line mechanism with proper context branching for
constrained strong and constrained marked.

In order to do so, the `prototype` context was also added to syntax,
in order to handle (single) backlash escaping and XML entities.

Now the syntax can properly handle most (if not all) cases of literal
hash and asterisk symbols within the text, by gracefully failing to
an alternative context branch which treats them as literal characters.
(it currently also scopes them as such, for testing purposes).

Context branching is a new Sublime Text 4 feature which wasn't available
at the time the original package was written (ST2), so this is the first
milestone in a series of syntax refactoring operations aimed at improving
the syntax to handle complex AsciiDoc constructs.

New tests were added and previous one were updated, properly documenting
the new feature.