Fix Callouts Test
Amend the new syntax test for callout lists by providing consecutive
callout numbers, so that Asciidoctor can build the HTML document
without error.

Create a copy of the original test (disjointed numbering sequence)
with `.adoc` extension, so that it will be ignored by the Rake task
that builds the HTML docs from the test files.