Build system for AsciiDoc files that targets HTML, EPUB, and DOCX. Fixes https://github.com/tajmone/ST4-Asciidoctor/issues/8
For HTML, this assumes that ASCIIDOCTOR is installed on the system path.
For EPUB, this assumes that ASCIIDOCTOR-EPUB3 is also installed on the system path.
For DOCX, this assumes that PANDOC is also installed on the system path.

NOTE: There is currently only a WINDOWS implementation for DOCX. (Someone just needs to translate the BAT file into SH scripts for Linux and Mac -- I'll open a pair of new issues for that.)