Add Goto Features for block IDs
1. Changed the block ID (anchor) scope from "markup.underline.blockid.id" to "entity.name.label" (per the syntax naming guidelines).
Added a preference P-list for the block ID (entity.name.label) to be indexed and thus appear in the symbol list (for declaration goto targeting).
Likewise, added a preference P-list for the block id within a link (variable.parameter.xref) to be indexed and thus appear in the symbol list (for reference goto targeting).

2. Removed the extraneous "level" from "markup.heading.level.1.asciidoc", "markup.heading.level.2.asciidoc", etc. to align with what the color schemes expect (e.g. the Monokai and Breakers schemes set one color for all "markup.heading" and a second color for "markup.heading.1" in particular, and Celeste also sets a third color for "markup.heading.2").